move throttling rate files into their own directory
