+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/settings/manage/throttlingRate/ThrottlingRate.tsx
