+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/settings/manage/throttlingRate/actions.ts
@@ -10,8 +10,6 @@ import { AuthenticatedAction } from "@lib/actions";
 import { ServerActionError } from "@lib/types/form-builder-types";
 import { logEvent } from "@lib/auditLogs";
 
-// Move to a separate ThrottlingRate directory. See https://github.com/cds-snc/platform-forms-client/issues/5814
-
 // Public facing functions - they can be used by anyone who finds the associated server action identifer
 
 export const getCurrentThrottlingRate = AuthenticatedAction(async (_, formId: string) => {
